python/observation-publisher/requirements.txt
confluent-kafka>=2.3.0
caraml-upi-protos>=1.0.0
arize==7.7.*
hydra-core>=1.3.0
pandas>=1.0.0

file:../sdk
